requirements.txt
# Core dependencies
flask>=3.1.2
flask-cors>=6.0.1
groq>=0.37.1
python-dotenv>=0.21.0

# Production WSGI server
gunicorn>=22.0.0

# Security and stability
werkzeug>=3.0.0
